Move IMappingCollector Related stuff to extra package
