+++ src/main/java/moze_intel/projecte/emc/IValueGenerator.java
@@ -1,5 +1,7 @@
 package moze_intel.projecte.emc;
 
+import moze_intel.projecte.emc.collector.IMappingCollector;
+
 import java.util.Map;
 
 public interface IValueGenerator<T, V extends Comparable<V>> extends IMappingCollector<T, V>
