+++ src/main/java/moze_intel/projecte/emc/mappers/APICustomEMCMapper.java
@@ -1,7 +1,7 @@
 package moze_intel.projecte.emc.mappers;
 
 import moze_intel.projecte.api.ProjectEAPI;
-import moze_intel.projecte.emc.IMappingCollector;
+import moze_intel.projecte.emc.collector.IMappingCollector;
 import moze_intel.projecte.emc.NormalizedSimpleStack;
 import moze_intel.projecte.utils.PELogger;
